feat: verify and save openai key
